initial commit
There's a simulator that allows you to test out various scenarios. It
relies on weighted random sampling from the Pareto distribution which
increases in latency depending on simulated concurrent requests. I'm not
100% sure about the formulate for the latter but it seems generally
reasonable.

The output? So far, I'm seeing resonance effects. I might not have
experimented enough with the parameters, though. My theory is that I
should modify the controller to not react on every sample, but rather
maybe an average of them, or possible a subsample of them.

Stuff that's remaining:

 * Renaming some of the types and functions. I think for example
   "Orchestrator" likely should be renamed to "WorkerPool". It makes so
   much more sense.
 * A bunch of small TODOs sprinkled across the code.